Manages my DNS records with OctoDNS
TL;DR
-----

Prepares the environment and data for using OctoDNS

Details
-------

Provides an initial implementation of OctoDNS for my public-facing DNS
records. This starting point allows me to start manging DNS records as
code again and is a foundation for maintaining records for all three
of my DNS horizons.

I started in on this since I was starting to expose a lot of services
from the lab via Tailscale and needed to set some public DNS records
for them. I took the opporunity to extend that to my other public
domains or split-horizon domains.